Move `aligned_size_t` and `get_device_address` to `cuda/__memory/`
