+++ docs/libcudacxx/extended_api/functional.rst
@@ -8,7 +8,7 @@ Functional
    :maxdepth: 1
 
    functional/proclaim_return_type
-   functional/get_device_address
+   memory/get_device_address
    functional/maximum_minimum
 
 .. list-table::
@@ -40,7 +40,7 @@ Functional
      - CCCL 2.8.0
      - CUDA 12.9
 
-   * - :ref:`cuda::get_device_address <libcudacxx-extended-api-functional-get-device-address>`
+   * - :ref:`cuda::get_device_address <libcudacxx-extended-api-memory-get-device-address>`
      - Returns a valid address to a device object
      - CCCL 2.8.0
      - CUDA 12.9
