+++ docs/libcudacxx/extended_api/memory.rst
@@ -7,6 +7,8 @@ Memory
    :hidden:
    :maxdepth: 1
 
+   memory/get_device_address
+   memory/aligned_size
    memory/is_aligned
    memory/align_up
    memory/align_down
@@ -20,6 +22,16 @@ Memory
      - **Content**
      - **CCCL Availability**
      - **CUDA Toolkit Availability**
+
+   * - :ref:`aligned_size_t <libcudacxx-extended-api-memory-aligned-size>`
+     - Defines an extent of bytes with a statically defined alignment.
+     - libcu++ 1.2.0 / CCCL 2.0.0
+     - CUDA 11.1
+
+   * - :ref:`get_device_address <libcudacxx-extended-api-memory-get-device-address>`
+     - Returns a valid address to a device object
+     - CCCL 2.8.0
+     - CUDA 12.9
 
    * - :ref:`is_aligned <libcudacxx-extended-api-memory-is_aligned>`
      - Check if a pointer is aligned
