--- docs/libcudacxx/extended_api/shapes.rst
@@ -1,29 +1,0 @@
-.. _libcudacxx-extended-api-memory-access-shapes:
-
-Shapes
-======
-
-.. toctree::
-   :hidden:
-   :maxdepth: 1
-
-   shapes/aligned_size_t
-
-.. list-table::
-   :widths: 25 45 30 30
-   :header-rows: 1
-
-   * - **Header**
-     - **Content**
-     - **CCCL Availability**
-     - **CUDA Toolkit Availability**
-
-   * - `cuda::std::size_t <https://en.cppreference.com/w/cpp/types/size_t>`_
-     - Defines an extent of bytes
-     - libcu++ 1.0.0 / CCCL 2.0.0
-     - CUDA 10.2
-
-   * - :ref:`cuda::aligned_size_t <libcudacxx-extended-api-memory-access-shapes-aligned-size>`
-     - Defines an extent of bytes with a statically defined alignment.
-     - libcu++ 1.2.0 / CCCL 2.0.0
-     - CUDA 11.1
