+++ libcudacxx/include/cuda/__memcpy_async/memcpy_async_tx.h
@@ -25,11 +25,11 @@
 #if _CCCL_CUDA_COMPILATION()
 #  if __cccl_ptx_isa >= 800
 
-#    include <cuda/__barrier/aligned_size.h>
 #    include <cuda/__barrier/async_contract_fulfillment.h>
 #    include <cuda/__barrier/barrier_block_scope.h>
 #    include <cuda/__barrier/barrier_native_handle.h>
 #    include <cuda/__memcpy_async/check_preconditions.h>
+#    include <cuda/__memory/aligned_size.h>
 #    include <cuda/__ptx/instructions/cp_async_bulk.h>
 #    include <cuda/__ptx/ptx_dot_variants.h>
 #    include <cuda/__ptx/ptx_helper_functions.h>
