+++ libcudacxx/include/cuda/memory
@@ -21,10 +21,11 @@
 #  pragma system_header
 #endif // no system header
 
-#include <cuda/__functional/get_device_address.h>
 #include <cuda/__memory/address_space.h>
 #include <cuda/__memory/align_down.h>
 #include <cuda/__memory/align_up.h>
+#include <cuda/__memory/aligned_size.h>
+#include <cuda/__memory/get_device_address.h>
 #include <cuda/__memory/is_aligned.h>
 #include <cuda/__memory/ptr_rebind.h>
 #include <cuda/std/memory>
